Add test or prod version

+++ bin/add_genome.sh
@@ -1,17 +1,22 @@
 set -e
 
 function usage(){
-    echo "usage : add_genome.sh -f=<path> --taxid=<int>|--taxname=<str> --id=<str> [-o=<path>]
-	[OPTIONS]
-	-h : print help
-	-f=<path> : path to fasta genome to add in database
-    -o=<path> : output directory to write files to insert (default : .)
-    --taxid=<int> : ncbi taxid (you can't combine --taxid and --taxname)
-    --id=<str> : genome identifiant (for example, GCF_xxxxxxx for Refseq genome)
-    --motif-broker-url=<url> : motif broker listening url (defaut : http://localhost:3282)
-    --taxonDB_name=<str> : taxon_db database name (default : taxon_db)
-    --db_url=<url> : database url (default : http://localhost:5984)
-    --taxonTree_DBname : taxon_tree database name (default : taxon_tree)
+    echo "usage : add_genome.sh -f <path> --taxid=<int> --id=<str>
+    [OPTIONS]
+    -h : print help
+    [Inputs]
+    -f <path> : path to fasta genome to add in database
+    --taxid <int> : ncbi taxid
+    --id <str> : genome identifiant (for example, GCF_xxxxxxx for Refseq genome)
+    [Outputs]
+    -o <path> : output directory to write files to insert (default : .)
+    [Databases]
+    --motif-broker-url <url> : motif broker listening url (defaut : http://localhost:3282)
+    --db_url <url> : database url (default : http://localhost:1234)
+    --taxonDB_name <str> : taxon_db database name (default : taxon_db)
+    --taxonTree_name <str> : taxon_tree database name (default : taxon_tree)
+    [Others]
+    --test : test version, database url will be http://localhost:5984 and files will not be copied to arwen and arwen-dev
     "
 }
 
@@ -43,11 +48,15 @@ function args_gestion(){
     fi
 
     if [[ ! $DB_URL ]]; then
-        DB_URL="http://localhost:5984/"
+        DB_URL="http://localhost:1234/"
     fi
 
     if [[ ! $TREE_DB_NAME ]]; then
         TREE_DB_NAME="taxon_tree"
+    fi
+
+    if [[ $TEST ]]; then
+        DB_URL="http://localhost:5984"
     fi
 
     if [[ $quit ]]; then
@@ -65,48 +74,59 @@ function args_verif(){
         mkdir -p $OUTDIR
     fi
 
+    #Verif arwen-dev is mount
+    if [[ ! $TEST ]]; then
+        if [[ ! $(ls /mnt/arwen-dev) ]]; then
+            echo "arwen-dev is not mount"
+            exit
+        fi
+
+        if [[ ! $(ls /mnt/arwen) ]]; then
+            echo "arwen is not mount"
+            exit
+        fi
+    fi
 }
 
+TEMP=$(getopt -o h,o:,f: -l taxid:,id:,motif-broker-url:,db_url:,taxonDB_name:,taxonTree_name:,test -- "$@")
 
-while [ "$1" != "" ]; do
-    PARAM=`echo $1 | awk -F= '{print $1}'`
-    VALUE=`echo $1 | awk -F= '{print $2}'`
-    case $PARAM in
-        -h | --help)
-            usage
-            exit
-            ;;
-        -f)
-            FASTA=$VALUE
-            ;;
-        -o)
-            OUTDIR=$VALUE
-            ;;
+eval set -- "$TEMP"
+
+while true ; do 
+	case "$1" in 
+		-f) 
+			FASTA=$2
+			shift 2;; 
+		-o) 
+			OUTDIR=$2
+			shift 2;; 
         --taxid)
-            TAXID=$VALUE
-            ;;      
+            TAXID=$2
+            shift 2;; 
         --id)
-            GENOME_ID=$VALUE
-            ;;
+            GENOME_ID=$2
+            shift 2;;
         --motif-broker-url)
-            MB_URL=$VALUE
-            ;;  
+            MB_URL=$2
+            shift 2;; 
         --taxonDB_name)
-            TAXONDB_NAME=$VALUE
-            ;;  
+            TAXONDB_NAME=$2
+            shift 2;;
         --db_url)
-            DB_URL=$VALUE
-            ;;
-        --taxonTree_DBname)
-            TREE_DB_NAME=$VALUE
-            ;;
-        *)
-            echo "ERROR: unknown parameter \"$PARAM\""
-            usage
-            exit 1
-            ;;
-    esac
-    shift
+            DB_URL=$2
+            shift 2;;
+        --taxonTree_name)
+            TREE_DB_NAME=$2
+            shift 2;;
+        --test)
+            TEST=1
+            shift;;   
+		-h) 
+			usage 
+			shift ;;
+		--)  
+			shift ; break ;; 					
+	esac 
 done
 
 args_gestion
@@ -121,49 +141,67 @@ elif [[ $tool_dir == $0 ]]; then
 fi 
 BIN=$tool_dir/bin
 
-#Verif arwen-dev is mount
-if [[ ! $(ls /mnt/arwen-dev) ]]; then
-    echo "arwen-dev is not mount"
-    exit
+TIMESTAMP=$(date +%Y%m%d%H%M%S)
+RESDIR=$OUTDIR/add_genome_$TIMESTAMP
+mkdir -p $RESDIR
+
+echo " Fasta file : $FASTA
+Taxid : $TAXID
+Id : $ID
+Output directory : $RESDIR
+Database url : $DB_URL
+Taxon database name : $TAXONDB_NAME
+Taxon tree database name : $TREE_DB_NAME
+motif-broker end point : $MB_URL 
+"
+if [[ $TEST ]];then
+    echo "TEST VERSION"
+else 
+    echo "PRODUCTION VERSION"
 fi
 
-if [[ ! $(ls /mnt/arwen) ]]; then
-    echo "arwen is not mount"
+read -p "Continue ? (y/n)" cont
+while [[ $cont != "y" && $cont != "n" ]]; do
+    read -p "Continue ? (y/n)" cont
+done
+
+if [[ $cont == "n" ]]; then
     exit
 fi
-
-RESDIR=$(readlink -f $(mktemp -d -p $OUTDIR))
-echo RESDIR $RESDIR
-
-echo "== Search sgRNA, indexing genome"
-# TO DO : parallelize construction
-mkdir -p $RESDIR/genome_pickle
-mkdir -p $RESDIR/genome_index
-echo Launch python $BIN/create_metafile.py -file "$FASTA" -rfg "$RESDIR" -taxid "$TAXID" -gcf "$GENOME_ID"
-python $BIN/create_metafile.py -file "$FASTA" -rfg "$RESDIR" -taxid "$TAXID" -gcf "$GENOME_ID"
-
-echo "== Add sgRNA to couchDB"
-echo Launch python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py --map /home/chilpert/app/motif-broker-2/data/3letter_prefixe_rules.json --data "$RESDIR/genome_pickle" --url "$DB_URL"
-python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py --map /home/chilpert/app/motif-broker-2/data/3letter_prefixe_rules.json --data "$RESDIR/genome_pickle" --url "$DB_URL"
-
-echo "== Copy index to arwen-dev"
-echo Launch cp $RESDIR/genome_index/* /mnt/arwen-dev/data/databases/mobi/crispr_clean/genome_index/
-cp $RESDIR/genome_index/* /mnt/arwen-dev/data/databases/mobi/crispr_clean/genome_index/
-
-echo "== Copy pickle to arwen"
-cp $RESDIR/genome_pickle/* /mnt/arwen/mobi/group/databases/crispr/crispr_rc01/pickle/
 
 echo "== Add to taxon_db"
 echo Launch python $BIN/create_file_taxondb.py single -gcf "$GENOME_ID" -taxid "$TAXID" -r "$MB_URL" -dbName "$TAXONDB_NAME" -fasta "$FASTA" -outdir "$RESDIR"
 python $BIN/create_file_taxondb.py single -gcf "$GENOME_ID" -taxid "$TAXID" -r "$MB_URL" -dbName "$TAXONDB_NAME" -fasta "$FASTA" -outdir "$RESDIR"
 
+echo "== Search sgRNA, indexing genome"
+mkdir -p $RESDIR/genome_pickle
+mkdir -p $RESDIR/genome_index
+echo Launch python $BIN/create_metafile.py -file "$FASTA" -rfg "$RESDIR" -taxid "$TAXID" -gcf "$GENOME_ID"
+python $BIN/create_metafile.py -file "$FASTA" -rfg "$RESDIR" -taxid "$TAXID" -gcf "$GENOME_ID"
+
+echo "== Insert sgRNA in database"
+echo Launch python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py --map /home/chilpert/app/motif-broker-2/data/3letter_prefixe_rules.json --data "$RESDIR/genome_pickle" --url "$DB_URL"
+python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py --map /home/chilpert/app/motif-broker-2/data/3letter_prefixe_rules.json --data "$RESDIR/genome_pickle" --url "$DB_URL"
+
+if [[ ! $TEST ]];then
+    echo "== Copy index to arwen-dev"
+    echo Launch cp $RESDIR/genome_index/* /mnt/arwen-dev/data/databases/mobi/crispr_clean/genome_index/
+    cp $RESDIR/genome_index/* /mnt/arwen-dev/data/databases/mobi/crispr_clean/genome_index/
+
+    echo "== Copy pickle to arwen"
+    echo Launch cp $RESDIR/genome_pickle/* /mnt/arwen/mobi/group/databases/crispr/crispr_rc01/pickle/
+    cp $RESDIR/genome_pickle/* /mnt/arwen/mobi/group/databases/crispr/crispr_rc01/pickle/
+
+fi 
+
+echo "Insert taxon in database"
 echo Launch python ~/Dev/pyCouch/scripts/couchBuild.py taxon_db --url "$DB_URL" --data $RESDIR/taxonDB_data
 python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py taxon_db --url "$DB_URL" --data $RESDIR/taxonDB_data
 
 echo "== Recalculate tree"
-echo Launch python $BIN/create_tree.py -url "$DB_URL" -taxonDB_name "$TAXONDB_NAME" -o "$RESDIR"
-python $BIN/create_tree.py -url "$DB_URL" -taxonDB_name "$TAXONDB_NAME" -o "$RESDIR"
+echo Launch python $BIN/create_tree.py -url "$DB_URL/" -taxonDB_name "$TAXONDB_NAME" -o "$RESDIR"
+python $BIN/create_tree.py -url "$DB_URL/" -taxonDB_name "$TAXONDB_NAME" -o "$RESDIR"
 
-echo "==Insert tree"
+echo "==Insert tree in database"
 echo Launch python ~/Dev/pyCouch/scripts/couchBuild.py $TREE_DB_NAME --url $DB_URL --data $RESDIR/treeDB_data
 python /home/chilpert/Dev/pyCouch/scripts/couchBuild.py $TREE_DB_NAME --url $DB_URL --data $RESDIR/treeDB_data
